added smoothed and ply file reading

+++ raytracer/BRDF/BRDF.hpp
@@ -24,6 +24,6 @@ class BRDF{
         virtual ~BRDF() = default;
 
         virtual RGBColor f(const ShadeInfo& sr, const Vector3D& wi, const Vector3D& wo) const = 0;
-        virtual RGBColor sample_f(const ShadeInfo& sr, Vector3D& wi, const Vector3D& wo, float& pdf) const = 0;
+        virtual RGBColor sample_f(const ShadeInfo& sr, Vector3D& wi, const Vector3D& wo, double& pdf) const = 0;
         virtual RGBColor rho(const ShadeInfo& sr, const Vector3D& wo) const = 0;
 };
