+++ raytracer/BRDF/GlossySpecular.cpp
@@ -1,7 +1,7 @@
 #include "GlossySpecular.hpp"
 #include <math.h>
 
-GlossySpecular::GlossySpecular(float ks_, RGBColor cs_, float exp_){
+GlossySpecular::GlossySpecular(double ks_, RGBColor cs_, double exp_){
     ks = ks_;
     cs = cs_;
     exp = exp_;
@@ -9,9 +9,9 @@ GlossySpecular::GlossySpecular(float ks_, RGBColor cs_, float exp_){
 
 RGBColor GlossySpecular::f(const ShadeInfo& sr, const Vector3D& wi, const Vector3D& wo) const {
     RGBColor L; 
-    float ndotwi = sr.normal * wi;
+    double ndotwi = sr.normal * wi;
     Vector3D r(-wi + 2.0 * sr.normal * ndotwi);
-    float rdotwo = r * wo;
+    double rdotwo = r * wo;
 
     if (rdotwo > 0.0)
         L = ks * pow(rdotwo, exp);
@@ -19,8 +19,8 @@ RGBColor GlossySpecular::f(const ShadeInfo& sr, const Vector3D& wi, const Vector
     return (L);
 }
 
-RGBColor GlossySpecular::sample_f(const ShadeInfo& sr, Vector3D& wi, const Vector3D& wo, float& pdf) const {
-    float ndotwo = sr.normal * wo;
+RGBColor GlossySpecular::sample_f(const ShadeInfo& sr, Vector3D& wi, const Vector3D& wo, double& pdf) const {
+    double ndotwo = sr.normal * wo;
     Vector3D r = -wo + 2.0 * sr.normal * ndotwo; // direction of mirror
                                                     // reflection
     Vector3D w = r;
@@ -34,7 +34,7 @@ RGBColor GlossySpecular::sample_f(const ShadeInfo& sr, Vector3D& wi, const Vecto
     if (sr.normal * wi < 0.0) // reflected ray is below surface
         wi = -sp.x * u - sp.y * v + sp.z * w;
 
-    float phong_lobe = pow(r * wi, exp);
+    double phong_lobe = pow(r * wi, exp);
     pdf = phong_lobe * (sr.normal * wi);
 
     return (ks * cs * phong_lobe);
@@ -44,7 +44,7 @@ RGBColor GlossySpecular::rho(const ShadeInfo& sr, const Vector3D& wo) const{
     return RGBColor(); // return black color
 }
 
-void GlossySpecular::set_ks(const float _ks){
+void GlossySpecular::set_ks(const double _ks){
     ks = _ks;
 }
 
@@ -52,11 +52,11 @@ void GlossySpecular::set_cs(const RGBColor& _cs){
     cs = _cs;
 }
 
-void GlossySpecular::set_exp(const float _exp){
+void GlossySpecular::set_exp(const double _exp){
     exp = _exp;
 }
 
-float GlossySpecular::get_ks() const{
+double GlossySpecular::get_ks() const{
     return ks;
 }
