+++ raytracer/BRDF/Lambertian.cpp
@@ -8,11 +8,11 @@ RGBColor Lambertian::rho(const ShadeInfo& sr, const Vector3D& wo) const{
     return (kd * cd);
 }
 
-RGBColor Lambertian::sample_f(const ShadeInfo& sr, Vector3D& wi, const Vector3D& wo, float& pdf){
+RGBColor Lambertian::sample_f(const ShadeInfo& sr, Vector3D& wi, const Vector3D& wo, double& pdf){
     return RGBColor();
 }
 
-void Lambertian::set_kd(const float _kd){
+void Lambertian::set_kd(const double _kd){
     kd = _kd;
 }
 
@@ -20,7 +20,7 @@ void Lambertian::set_cd(const RGBColor& _cd){
     cd = _cd;
 }
 
-float Lambertian::get_kd() const{
+double Lambertian::get_kd() const{
     return kd;
 }
 
@@ -28,7 +28,7 @@ RGBColor Lambertian::get_cd() const{
     return cd;
 }
 
-Lambertian::Lambertian(float kd_, RGBColor cd_){
+Lambertian::Lambertian(double kd_, RGBColor cd_){
     kd = kd_;
     cd = cd_;
 }
