+++ raytracer/BRDF/PerfectSpecular.cpp
@@ -1,13 +1,13 @@
 #include "PerfectSpecular.hpp"
 
 // Constructor
-PerfectSpecular::PerfectSpecular(float kr_, RGBColor cr_){
+PerfectSpecular::PerfectSpecular(double kr_, RGBColor cr_){
     kr = kr_;
     cr = cr_;
 }
 
 RGBColor PerfectSpecular::sample_f(const ShadeInfo& sr, Vector3D& wi, const Vector3D& wo) const {
-    float ndotwo = sr.normal * wo;
+    double ndotwo = sr.normal * wo;
     wi = -wo + 2.0 * sr.normal * ndotwo;
     return (kr * cr / (sr.normal * wi));
 }
@@ -20,7 +20,7 @@ RGBColor PerfectSpecular::rho(const ShadeInfo& sr, const Vector3D& wo) const{
     return RGBColor(); // return black color
 }
 
-void PerfectSpecular::set_kr(const float _kr){
+void PerfectSpecular::set_kr(const double _kr){
     kr = _kr;
 }
 
@@ -28,7 +28,7 @@ void PerfectSpecular::set_cr(const RGBColor& _cr){
     cr = _cr;
 }
 
-float PerfectSpecular::get_kr() const{
+double PerfectSpecular::get_kr() const{
     return kr;
 }
