+++ raytracer/BRDF/PerfectSpecular.hpp
@@ -5,14 +5,14 @@
 class PerfectSpecular{
     
     private:
-        float kr;
+        double kr;
         RGBColor cr;
     
     public:
 
         // Constructors.
         PerfectSpecular() = default;
-        PerfectSpecular(float kr_, RGBColor cr_);
+        PerfectSpecular(double kr_, RGBColor cr_);
 
         // Destructor.
         ~PerfectSpecular() = default;
@@ -22,8 +22,8 @@ class PerfectSpecular{
         virtual RGBColor rho(const ShadeInfo& sr, const Vector3D& wo) const;
 
         // Access Methods.
-        void set_kr(const float _kr);
+        void set_kr(const double _kr);
         void set_cr(const RGBColor& _cr);
-        float get_kr() const;
+        double get_kr() const;
         RGBColor get_cr() const;
 };
