+++ raytracer/cameras/Parallel.cpp
@@ -4,12 +4,12 @@ Parallel::Parallel(){
     dir = Vector3D(0, 0, -1);
 }
 
-Parallel::Parallel(float c){
+Parallel::Parallel(double c){
     dir = Vector3D(c,c,c);
     dir.normalize();
 }
 
-Parallel::Parallel(float x, float y, float z){
+Parallel::Parallel(double x, double y, double z){
     dir = Vector3D(x,y,z);
     dir.normalize();
 }
