+++ raytracer/cameras/Parallel.hpp
@@ -18,8 +18,8 @@ protected:
 public:
   // Constructors.
   Parallel();        // set dir parallel to -z (negative z) axis.
-  Parallel(float c); // set dir parallel to (c, c, c).
-  Parallel(float x, float y, float z); // set dir parallel to (x, y, z)
+  Parallel(double c); // set dir parallel to (c, c, c).
+  Parallel(double x, double y, double z); // set dir parallel to (x, y, z)
   Parallel(const Vector3D &d);         // set dir parallel to d.
 
   // Copy constuctor and assignment operator.
