+++ raytracer/lights/AmbientLight.cpp
@@ -3,7 +3,7 @@
 AmbientLight::AmbientLight(): ls(1), color(1){
 }
 
-AmbientLight::AmbientLight(float l, RGBColor col): ls(l), color(col){
+AmbientLight::AmbientLight(double l, RGBColor col): ls(l), color(col){
 }
 
 Vector3D AmbientLight::get_direction(ShadeInfo &si){
