+++ raytracer/lights/PointLight.cpp
@@ -3,7 +3,7 @@
 PointLight::PointLight(): ls(1), color(1), location(){
 }
 
-PointLight::PointLight(float l, RGBColor col, Vector3D loc): ls(l), color(col), location(loc){
+PointLight::PointLight(double l, RGBColor col, Vector3D loc): ls(l), color(col), location(loc){
 }
 
 Vector3D PointLight::get_direction(ShadeInfo &si){
@@ -14,10 +14,10 @@ RGBColor PointLight::L(ShadeInfo &si) {
     return (ls * color);
 }
 
-void PointLight::set_location(const float x, const float y, const float z){
+void PointLight::set_location(const double x, const double y, const double z){
     location = Vector3D(x,y,z);
 }
 
-void PointLight::scale_radiance(const float f){
+void PointLight::scale_radiance(const double f){
     ls = f;
 }
