+++ raytracer/samplers/Sampler.cpp
@@ -62,19 +62,19 @@ Point3D Sampler::sample_hemisphere(){
     return hemisphere_samples[jump + shuffled_indices[jump + count++ % num_samples]];
 }
 
-void Sampler::map_samples_to_hemisphere(const float exp){
+void Sampler::map_samples_to_hemisphere(const double exp){
 
     int size = samples.size();
 	hemisphere_samples.reserve(num_samples * num_sets);
 		
 	for (int j = 0; j < size; j++) {
-		float cos_phi = cos(2.0 * PI * samples[j].x);
-		float sin_phi = sin(2.0 * PI * samples[j].x);	
-		float cos_theta = pow((1.0 - samples[j].y), 1.0 / (exp + 1.0));
-		float sin_theta = sqrt (1.0 - cos_theta * cos_theta);
-		float pu = sin_theta * cos_phi;
-		float pv = sin_theta * sin_phi;
-		float pw = cos_theta;
+		double cos_phi = cos(2.0 * PI * samples[j].x);
+		double sin_phi = sin(2.0 * PI * samples[j].x);	
+		double cos_theta = pow((1.0 - samples[j].y), 1.0 / (exp + 1.0));
+		double sin_theta = sqrt (1.0 - cos_theta * cos_theta);
+		double pu = sin_theta * cos_phi;
+		double pv = sin_theta * sin_phi;
+		double pw = cos_theta;
 		hemisphere_samples.push_back(Point3D(pu, pv, pw)); 
 	}
 }
