+++ raytracer/utilities/Ray.hpp
@@ -13,7 +13,7 @@ class Ray {
 public:
   Point3D o;  // origin
   Vector3D d; // direction, keep normalized.
-  float w;    // weightage of ray for a particular pixel, default is 1.
+  double w;    // weightage of ray for a particular pixel, default is 1.
 
 public:
   // Constructors.
